Remove js import cycles
I see in cloudflare logs that there is an exception about importing
undefined and a quick search reveals that this is often due to import
cycles (which posthog.ts and logging.ts) have.

This PR fixes up all import cycles and denies new ones via oxlint.
