+++ .oxlintrc.json
@@ -3,7 +3,8 @@
     // Consistent type imports is critical to avoid accidentally importing code
     // referencing window in a web worker
     "typescript/consistent-type-imports": "error",
-    "import/consistent-type-specifier-style": "error"
+    "import/consistent-type-specifier-style": "error",
+    "import/no-cycle": "error"
   },
   "categories": {
     "correctness": "error"
