+++ src/app/app/features/eu4/store/eu4Store.tsx
@@ -14,7 +14,7 @@ import type {
   MapDate,
   CountryTag,
 } from "../types/models";
-import { getEu4Worker } from "../worker";
+import { getEu4Worker } from "../worker/getEu4Worker";
 import type {
   EnhancedMeta,
   FileObservationFrequency,
