+++ src/map/src/index.ts
@@ -1,5 +1,5 @@
 export { GLResources } from "./glResources";
-export { IMG_HEIGHT, IMG_WIDTH, WebGLMap, glContextOptions } from "./map";
+export { WebGLMap, glContextOptions } from "./map";
 export { MapShader } from "./MapShader";
 export { ProvinceFinder } from "./ProvinceFinder";
 export { compileShaders } from "./shaderCompiler";
@@ -10,6 +10,7 @@ export { type InitToken } from "./map-worker";
 export * from "./canvasOverlays";
 export * from "./map-worker-types";
 export { provinceIdToColorIndexInvert } from "./resources";
+export { IMG_HEIGHT, IMG_WIDTH } from "./mapDimensions";
 
 export function createMapWorker() {
   return new Worker(new URL("./map-worker-bridge", import.meta.url), {
