+++ src/map/src/mapDimensions.ts
@@ -1,0 +1,4 @@
+export const IMG_HEIGHT = 2048;
+export const IMG_WIDTH = 5632;
+export const IMG_PADDED_WIDTH = 8192;
+export const SPLIT_IMG_PADDED_WIDTH = 4096;
